fixed code organisation
